additional development

+++ CCC.Data.Monitoring.Concrete/Entities/MonitorData.cs
@@ -7,10 +7,10 @@ namespace CCC.Data.Monitoring.Concrete.Entities
         public int ID { get; set; }
         public double TalkTime { get; set; }
         public double AfterCallWorkTime { get; set; }
-        public int Handled { get; set; }
-        public int Offered { get; set; }
-        public int HandledWithinSL { get; set; }
-        public int QueueGroupID { get; set; }
+        public double Handled { get; set; }
+        public double Offered { get; set; }
+        public double HandledWithinSL { get; set; }
+        public double QueueGroupID { get; set; }
 
     }
 }
