+++ CCC.Data.Monitoring.Concrete/Entities/ScreenData.cs
@@ -7,8 +7,8 @@ namespace CCC.Data.Monitoring.Concrete.Entities
     public class ScreenData
     {
         public string QueueGroupName { get; set; }
-        public int Offered { get; set; }
-        public int Handled { get; set; }
+        public double Offered { get; set; }
+        public double Handled { get; set; }
         public string AverageTalkTime { get; set; }
         public string AverageHandlingTime { get; set; }
         public string ServiceLevel { get; set; }
